preparing for silhouette

+++ conf/evolutions/default/2.sql
@@ -3,8 +3,8 @@
 # --- !Ups
 
 CREATE TABLE contacts (
-    id            BIGSERIAL     NOT NULL,
-    client_ID     INTEGER       NOT NULL,
+    id            BIGSERIAL       NOT NULL,
+    client_ID     BIGSERIAL       NOT NULL,
     mobile        VARCHAR(20)           ,
     description   VARCHAR(255)          ,
     created_at    TIMESTAMP WITH TIME ZONE NOT NULL,
